The performance of block allocator went down after implementing automatic prefix caching, even when running with prefix caching disabled. This pr brings back parts of the old code and regains some of the lost performance in the scenario with disabled prefix caching.

Benchmarked with:

`python benchmark_throughput_cache.py --backend vllm --model huggyllama/llama-7b --dataset ../data/ShareGPT_V3_unfiltered_cleaned_split.json --num-prompts 2000`

**Performance before introducing automatic prefix caching (commit baee28c46c242b72f90d6b1211ab9d7872ab05d3):**
Throughput: 10.37 requests/s, 5062.42 tokens/s
Throughput: 10.46 requests/s, 5102.27 tokens/s
Throughput: 10.47 requests/s, 5107.30 tokens/s
Throughput: 10.48 requests/s, 5113.97 tokens/s
Throughput: 10.53 requests/s, 5137.21 tokens/s
Throughput: 10.54 requests/s, 5145.38 tokens/s
Throughput: 10.56 requests/s, 5153.24 tokens/s
Throughput: 10.57 requests/s, 5157.54 tokens/s
Throughput: 10.63 requests/s, 5187.32 tokens/s
Throughput: 10.65 requests/s, 5198.19 tokens/s

**Performance after introducing changes in this PR to commit ce4f5a29fb3e35041842518fefe999847b8326b9:**
Throughput: 10.40 requests/s, 5076.05 tokens/s
Throughput: 10.53 requests/s, 5137.97 tokens/s
Throughput: 10.57 requests/s, 5156.04 tokens/s
Throughput: 10.60 requests/s, 5173.07 tokens/s
Throughput: 10.61 requests/s, 5177.02 tokens/s
Throughput: 10.62 requests/s, 5179.91 tokens/s
Throughput: 10.63 requests/s, 5186.06 tokens/s
Throughput: 10.63 requests/s, 5186.63 tokens/s
Throughput: 10.64 requests/s, 5193.72 tokens/s
Throughput: 10.67 requests/s, 5207.76 tokens/s

----------------------------------
(OLD)

Benchmark results (10 runs each):

**Performance before introducing automatic prefix caching (commit baee28c46c242b72f90d6b1211ab9d7872ab05d3):**
Throughput: 10.15 requests/s, 4909.50 tokens/s
Throughput: 10.17 requests/s, 4918.22 tokens/s
Throughput: 10.20 requests/s, 4936.93 tokens/s
Throughput: 10.23 requests/s, 4949.76 tokens/s
Throughput: 10.22 requests/s, 4945.64 tokens/s
Throughput: 10.27 requests/s, 4967.08 tokens/s
Throughput: 10.28 requests/s, 4971.52 tokens/s
Throughput: 10.29 requests/s, 4980.92 tokens/s
Throughput: 10.29 requests/s, 4976.94 tokens/s
Throughput: 10.30 requests/s, 4982.69 tokens/s

**Performance after introducing automatic prefix caching (commit ce4f5a29fb3e35041842518fefe999847b8326b9):**
Throughput: 9.91 requests/s, 4795.14 tokens/s
Throughput: 9.98 requests/s, 4830.01 tokens/s
Throughput: 9.99 requests/s, 4832.00 tokens/s
Throughput: 10.00 requests/s, 4839.62 tokens/s
Throughput: 10.03 requests/s, 4851.13 tokens/s
Throughput: 10.06 requests/s, 4868.87 tokens/s
Throughput: 10.07 requests/s, 4873.87 tokens/s
Throughput: 10.07 requests/s, 4872.51 tokens/s
Throughput: 10.08 requests/s, 4876.18 tokens/s
Throughput: 10.08 requests/s, 4877.26 tokens/s

**Performance after introducing changes in this PR to commit ce4f5a29fb3e35041842518fefe999847b8326b9:**
Throughput: 10.07 requests/s, 4873.42 tokens/s
Throughput: 10.17 requests/s, 4919.84 tokens/s
Throughput: 10.18 requests/s, 4923.71 tokens/s
Throughput: 10.18 requests/s, 4925.56 tokens/s
Throughput: 10.19 requests/s, 4928.09 tokens/s
Throughput: 10.20 requests/s, 4937.20 tokens/s
Throughput: 10.21 requests/s, 4942.21 tokens/s
Throughput: 10.21 requests/s, 4938.38 tokens/s
Throughput: 10.21 requests/s, 4940.22 tokens/s
Throughput: 10.22 requests/s, 4946.95 tokens/s